tonic: separate unary timing bench from allocation census
The counting allocator adds atomic operations to every timed RPC.
Move the allocation census to its own process so timing uses the
system allocator. Both targets use the same round-trip fixture.

Report allocation totals with per-RPC counts for each payload size.
