+++ tonic/Cargo.toml
@@ -146,3 +146,7 @@ name = "decode"
 [[bench]]
 name = "unary"
 harness = false
+
+[[bench]]
+harness = false
+name = "unary_allocs"
